Modified JQuery locators to be more precise
I've removed every first and lass CSS locators and replaced them with a eq expression. There is a hidden form on every page which is performing portal navigation and these locators were causing some tests to fail (which is obvious because expression last always filled field in navigation hidden form).